Create FontClass enum replacing string representation
The font class (by class reader should think of html, css class as down
on the rabbithole it kinda represents that) currently represented as
some sort of string: char*, QCString, ...

The valid values are limited to a finite set:
* no value (currently nullptr)
* keyword, comment, stringliteral, ...

Likely a proper refactoring steps would be:
* to extract *no value* into a std::optional (or something that acts like that).
* add new type that checks in constructor if the *string* is in the
  pre-defined list (convertion could help not to blow up the codebase)
* replace the new type representation with a simple enum

Instead I got lazy and did the reverse order, except in reverse order
the 2nd step is void.
Also there is a decision regarding what to do with *no value* if
std::option is not to be used yet. Well *none* value is added to be
removed later (using nullptr for nothing concidered worse).

That's it for now. Keep up the good work of reading the commit mesages in the next commit.

ps: sorry for the big patch, but it would have been more pain to deal
    with it in smaller chunks.

Signed-off-by: Kokan <kokaipeter@gmail.com>