Set c++ standard to 17 in order to access std::optional
The ratio, yes this probably makes the patch unmergable to upstream
(which is the worst thing ever), on the other hand standard lib writterr
had their heart and soult put into std::optional so I have the option to
save disk space and with that no more global warning (maybe there is).

If the c++17 is a no go for PR to be merged, one can simply write (yyp
if you know what I mean) a compat layer filling the gap. So this is just
making sure I do the task and not reinventing the wheel.

Note: std::optional might be better in the container.h this project
happens to have.

Signed-off-by: Kokan <kokaipeter@gmail.com>